Change project to work with array and not with reducer
