+++ src/HelpersFunc/WinCheckAndMark-object.mjs
